+++ lib/rodauth/rails/feature/internal_request.rb
@@ -4,14 +4,18 @@ module Rodauth
       module InternalRequest
         extend ActiveSupport::Concern
 
+        included do
+          auth_private_methods :rails_url_options
+        end
+
         def domain
-          return super unless missing_host? && rails_url_options
+          return super unless missing_host? && rails_url_options!
 
           rails_url_options.fetch(:host)
         end
 
         def base_url
-          return super unless missing_host? && domain && rails_url_options
+          return super unless missing_host? && domain && rails_url_options!
 
           scheme = rails_url_options[:protocol] || "http"
           port = rails_url_options[:port]
@@ -19,6 +23,11 @@ module Rodauth
           url = "#{scheme}://#{domain}"
           url << ":#{port}" if port
           url
+        end
+
+        def rails_url_options
+          return _rails_url_options if frozen? # handle path_class_methods feature
+          @rails_url_options ||= _rails_url_options
         end
 
         private
@@ -44,11 +53,12 @@ module Rodauth
           internal_request? && (request.host.nil? || request.host == INVALID_DOMAIN) || scope.nil?
         end
 
-        def rails_url_options
-          return nil unless defined?(ActionMailer)
+        def rails_url_options!
+          rails_url_options.presence or fail Error, "There is no information to set the URL host from. Please set config.action_mailer.default_url_options in your Rails application, configure #domain and #base_url in your Rodauth configuration, or set #rails_url_options on the Rodauth instance."
+        end
 
-          ::Rails.configuration.action_mailer.default_url_options or
-            fail Error, "There is no information to set the URL host from. Please set config.action_mailer.default_url_options in your Rails application, or configure #domain and #base_url in your Rodauth configuration."
+        def _rails_url_options
+          defined?(ActionMailer) ? ActionMailer::Base.default_url_options.dup : {}
         end
       end
     end